# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/leads": {
      "get": {
        "operationId": "LeadsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeadDetailResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "LeadsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/leads/cnpj/{cnpj}": {
      "get": {
        "operationId": "LeadsController_lookupCnpj",
        "parameters": [
          {
            "name": "cnpj",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/leads/{id}": {
      "get": {
        "operationId": "LeadsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "LeadsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLeadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "LeadsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/leads/{id}/contacts": {
      "post": {
        "operationId": "LeadsController_addContact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/leads/{id}/contacts/{contact_id}": {
      "patch": {
        "operationId": "LeadsController_updateContact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contact_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "LeadsController_removeContact",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contact_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/leads/{id}/comments": {
      "post": {
        "operationId": "LeadsController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/leads/{id}/comments/{comment_id}": {
      "patch": {
        "operationId": "LeadsController_updateComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "comment_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "LeadsController_removeComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "comment_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Leads"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente OU setor: comercial OU role: diretor E setor: marketing, comercial.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/push-subscriptions": {
      "post": {
        "operationId": "PushSubscriptionsController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePushSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "PushSubscriptions"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/push-subscriptions/{id}": {
      "delete": {
        "operationId": "PushSubscriptionsController_unregister",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "PushSubscriptions"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "WattAPI",
    "description": "Documentação gerada a partir dos schemas zod de request/response e do metadata de @RoutePolicy de cada rota.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LeadDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "company_name": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "address_logradouro": {
            "type": "string"
          },
          "address_numero": {
            "type": "string"
          },
          "address_complemento": {
            "type": "string",
            "nullable": true
          },
          "address_bairro": {
            "type": "string"
          },
          "address_cidade": {
            "type": "string"
          },
          "address_estado": {
            "type": "string"
          },
          "address_cep": {
            "type": "string"
          },
          "interest_items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "contacts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "lead_id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "email": {
                  "type": "string",
                  "nullable": true
                },
                "phone": {
                  "type": "string",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "lead_id",
                "name",
                "role",
                "email",
                "phone"
              ],
              "example": {
                "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "lead_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "name": "João Silva",
                "role": "Diretor",
                "email": "joao@empresa.com",
                "phone": null
              }
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "lead_id": {
                  "type": "string"
                },
                "user_id": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string"
                },
                "updated_at": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "lead_id",
                "user_id",
                "content",
                "created_at",
                "updated_at"
              ],
              "example": {
                "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "lead_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "user_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                "content": "Cliente demonstrou interesse em auditoria.",
                "created_at": "2026-05-25T12:00:00.000Z",
                "updated_at": "2026-05-25T12:00:00.000Z"
              }
            }
          }
        },
        "required": [
          "id",
          "company_name",
          "cnpj",
          "created_by",
          "status",
          "address_logradouro",
          "address_numero",
          "address_complemento",
          "address_bairro",
          "address_cidade",
          "address_estado",
          "address_cep",
          "interest_items",
          "created_at",
          "updated_at",
          "contacts",
          "comments"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "company_name": "Empresa ABC",
          "cnpj": "12.345.678/0001-95",
          "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "status": "nao_contatado",
          "address_logradouro": "Rua das Flores",
          "address_numero": "42",
          "address_complemento": null,
          "address_bairro": "Jardim Paulista",
          "address_cidade": "São Paulo",
          "address_estado": "SP",
          "address_cep": "01310100",
          "interest_items": [
            "Consultoria Energética"
          ],
          "contacts": [
            {
              "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "lead_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "name": "João Silva",
              "role": "Diretor",
              "email": "joao@empresa.com",
              "phone": null
            }
          ],
          "comments": [
            {
              "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "lead_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "user_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "content": "Cliente demonstrou interesse em auditoria.",
              "created_at": "2026-05-25T12:00:00.000Z",
              "updated_at": "2026-05-25T12:00:00.000Z"
            }
          ],
          "created_at": "2026-05-25T12:00:00.000Z",
          "updated_at": "2026-05-25T12:00:00.000Z"
        }
      },
      "CreateLeadDto": {
        "type": "object",
        "properties": {
          "company_name": {
            "type": "string",
            "minLength": 1
          },
          "cnpj": {
            "type": "string"
          },
          "address_logradouro": {
            "type": "string",
            "minLength": 1
          },
          "address_numero": {
            "type": "string",
            "minLength": 1
          },
          "address_complemento": {
            "type": "string"
          },
          "address_bairro": {
            "type": "string",
            "minLength": 1
          },
          "address_cidade": {
            "type": "string",
            "minLength": 1
          },
          "address_estado": {
            "type": "string",
            "minLength": 1
          },
          "address_cep": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "nao_contatado",
              "em_progresso",
              "contatado"
            ]
          },
          "interest_items": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "required": [
          "company_name",
          "cnpj",
          "address_logradouro",
          "address_numero",
          "address_bairro",
          "address_cidade",
          "address_estado",
          "address_cep"
        ],
        "example": {
          "company_name": "Empresa ABC",
          "cnpj": "12.345.678/0001-95",
          "address_logradouro": "Rua das Flores",
          "address_numero": "42",
          "address_bairro": "Jardim Paulista",
          "address_cidade": "São Paulo",
          "address_estado": "SP",
          "address_cep": "01310100",
          "status": "nao_contatado",
          "interest_items": [
            "Consultoria Energética"
          ]
        }
      },
      "LeadResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "company_name": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "created_by": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "address_logradouro": {
            "type": "string"
          },
          "address_numero": {
            "type": "string"
          },
          "address_complemento": {
            "type": "string",
            "nullable": true
          },
          "address_bairro": {
            "type": "string"
          },
          "address_cidade": {
            "type": "string"
          },
          "address_estado": {
            "type": "string"
          },
          "address_cep": {
            "type": "string"
          },
          "interest_items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "company_name",
          "cnpj",
          "created_by",
          "status",
          "address_logradouro",
          "address_numero",
          "address_complemento",
          "address_bairro",
          "address_cidade",
          "address_estado",
          "address_cep",
          "interest_items",
          "created_at",
          "updated_at"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "company_name": "Empresa ABC",
          "cnpj": "12.345.678/0001-95",
          "created_by": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "status": "nao_contatado",
          "address_logradouro": "Rua das Flores",
          "address_numero": "42",
          "address_complemento": null,
          "address_bairro": "Jardim Paulista",
          "address_cidade": "São Paulo",
          "address_estado": "SP",
          "address_cep": "01310100",
          "interest_items": [
            "Consultoria Energética"
          ],
          "created_at": "2026-05-25T12:00:00.000Z",
          "updated_at": "2026-05-25T12:00:00.000Z"
        }
      },
      "UpdateLeadDto": {
        "type": "object",
        "properties": {
          "company_name": {
            "type": "string",
            "minLength": 1
          },
          "cnpj": {
            "type": "string"
          },
          "address_logradouro": {
            "type": "string",
            "minLength": 1
          },
          "address_numero": {
            "type": "string",
            "minLength": 1
          },
          "address_complemento": {
            "type": "string"
          },
          "address_bairro": {
            "type": "string",
            "minLength": 1
          },
          "address_cidade": {
            "type": "string",
            "minLength": 1
          },
          "address_estado": {
            "type": "string",
            "minLength": 1
          },
          "address_cep": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "nao_contatado",
              "em_progresso",
              "contatado"
            ]
          },
          "interest_items": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          }
        },
        "example": {
          "company_name": "Empresa ABC Ltda",
          "status": "em_progresso"
        }
      },
      "CreateContactDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "role": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "phone": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name",
          "role"
        ],
        "example": {
          "name": "João Silva",
          "role": "Diretor",
          "email": "joao@empresa.com",
          "phone": "11999999999"
        }
      },
      "ContactResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "lead_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "lead_id",
          "name",
          "role",
          "email",
          "phone"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "lead_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "name": "João Silva",
          "role": "Diretor",
          "email": "joao@empresa.com",
          "phone": null
        }
      },
      "UpdateContactDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "role": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "minLength": 1,
            "nullable": true
          }
        },
        "example": {
          "phone": "11988888888"
        }
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "content"
        ],
        "example": {
          "content": "Cliente demonstrou interesse em auditoria."
        }
      },
...
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "content"
        ],
        "example": {
          "content": "Texto corrigido."
        }
      },
...
    }
  }
}